Update student dashboard & created select component (initial)
