+++ app/student/courses/page.tsx
@@ -1,15 +1,9 @@
 import React from "react";
 
 import CourseCard from "@/components/Student/CourseCardStudent";
-import {
-  Select,
-  SelectItem,
-  SelectContent,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
 
 import Search from "@/components/Student/Search";
+import CoursesSelect from "@/components/Student/CoursesSelect";
 
 // Fake data for demonstration
 const fakeCourses = [
@@ -88,66 +82,7 @@ const CoursesPage = async ({
           <div className="flex flex-1 items-center gap-2">
             <Search action="/student/courses" />
           </div>
-          <form className="flex flex-row gap-2" method="GET">
-            <div className="flex flex-col gap-2">
-              <div className="flex flex-row items-center gap-2">
-                <label htmlFor="sort" className="text-base-content/60 text-xs">
-                  Sorted by:
-                </label>
-              </div>
-              <Select name="sort">
-                <SelectTrigger className="w-40">
-                  <SelectValue placeholder="Sorted by" />
-                </SelectTrigger>
-                <SelectContent>
-                  <SelectItem value="Latest">Latest</SelectItem>
-                  <SelectItem value="Oldest">Oldest</SelectItem>
-                  <SelectItem value="Most Viewed">Most Viewed</SelectItem>
-                </SelectContent>
-              </Select>
-            </div>
-            <div className="flex flex-col gap-2">
-              <div className="flex flex-row items-center gap-2">
-                <label
-                  htmlFor="status"
-                  className="text-base-content/60 text-xs"
-                >
-                  Status:
-                </label>
-              </div>
-              <Select name="status">
-                <SelectTrigger className="w-40">
-                  <SelectValue placeholder="Status" />
-                </SelectTrigger>
-                <SelectContent>
-                  <SelectItem value="All Courses">All Courses</SelectItem>
-                  <SelectItem value="Ongoing">Ongoing</SelectItem>
-                  <SelectItem value="Completed">Completed</SelectItem>
-                </SelectContent>
-              </Select>
-            </div>
-            <div className="flex flex-col gap-2">
-              <div className="flex flex-row items-center gap-2">
-                <label
-                  htmlFor="teacher"
-                  className="text-base-content/60 text-xs"
-                >
-                  Teacher:
-                </label>
-              </div>
-              <Select name="teacher">
-                <SelectTrigger className="w-40">
-                  <SelectValue placeholder="Teacher" />
-                </SelectTrigger>
-                <SelectContent>
-                  <SelectItem value="All Teachers">All Teachers</SelectItem>
-                  <SelectItem value="Mr. Ahmadi">Mr. Ahmadi</SelectItem>
-                  <SelectItem value="Mr. Rezaei">Mr. Rezaei</SelectItem>
-                </SelectContent>
-              </Select>
-            </div>
-            <button type="submit" className="hidden" />
-          </form>
+          <CoursesSelect />
         </div>
       </div>
