+++ app/student/teachers/page.tsx
@@ -1,17 +1,9 @@
 import React from "react";
-import { Icon } from "@iconify/react";
-import Form from "next/form";
 
 import TeacherCard from "@/components/Student/TeacherCard";
-import {
-  Select,
-  SelectItem,
-  SelectContent,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
-import { redirect } from "next/navigation";
+
 import Search from "@/components/Student/Search";
+import TeacherSelect from "@/components/TeacherSelect";
 
 // Fake data for demonstration
 const fakeTeachers = [
@@ -69,50 +61,10 @@ const TeachersPage = ({
           <span className="text-base-content/80">{`(${filteredTeachers.length})`}</span>
         </div>
         <div className="flex flex-row gap-2">
-          {/* search */}
           <div className="flex flex-1 items-center gap-2">
-            {/* TODO: check the form */}
-
             <Search action="/student/teachers" />
           </div>
-          <Form className="flex flex-1 flex-row gap-2" action="">
-            <div className="flex flex-2 flex-col gap-2">
-              <div className="flex flex-row items-center gap-2">
-                <label htmlFor="sort" className="text-base-content/60 text-xs">
-                  Courses:
-                </label>
-              </div>
-              <Select name="courses">
-                <SelectTrigger className="w-40">
-                  <SelectValue placeholder="Courses" />
-                </SelectTrigger>
-                <SelectContent>
-                  <SelectItem value="All">All Courses</SelectItem>
-                  <SelectItem value="1">1 Course</SelectItem>
-                  <SelectItem value="2">2 Courses</SelectItem>
-                  <SelectItem value="3">3 Courses</SelectItem>
-                </SelectContent>
-              </Select>
-            </div>
-            <div className="flex flex-1 flex-col gap-2">
-              <div className="flex flex-row items-center gap-2">
-                <label htmlFor="sort" className="text-base-content/60 text-xs">
-                  Sort by:
-                </label>
-              </div>
-              <Select name="sort">
-                <SelectTrigger className="w-40">
-                  <SelectValue placeholder="Sort by" />
-                </SelectTrigger>
-                <SelectContent>
-                  <SelectItem value="Latest">Latest</SelectItem>
-                  <SelectItem value="Oldest">Oldest</SelectItem>
-                  <SelectItem value="Most Students">Most Students</SelectItem>
-                </SelectContent>
-              </Select>
-            </div>
-            <button type="submit" className="hidden" />
-          </Form>
+          <TeacherSelect />
         </div>
       </div>
       <div className="grid grid-cols-1 gap-6 sm:grid-cols-2 lg:grid-cols-4">
