+++ app/student/wishlist/page.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { Icon } from "@iconify/react";
-import Image from "next/image";
-import Link from "next/link";
 
+import Link from "next/link";
+import WishlistCourseRow from "@/components/Student/student-wishlist/WishlistCoursesRow";
 // Fake wishlist data for demonstration
 const fakeWishlistCourses = [
   {
@@ -82,114 +82,3 @@ export default function WishlistPage() {
 }
 
 // Custom Wishlist Course Row Component
-interface WishlistCourseRowProps {
-  id: number;
-  title: string;
-  image: string;
-  instructors: string;
-  price: string;
-  originalPrice: string | null;
-  rating: number;
-  reviews: number;
-}
-
-const WishlistCourseRow: React.FC<WishlistCourseRowProps> = ({
-  id,
-  title,
-  image,
-  instructors,
-  price,
-  originalPrice,
-  rating,
-  reviews,
-}) => {
-  return (
-    <div className="hover:bg-base-200/30 px-2 py-4 transition-colors">
-      <div className="grid grid-cols-12 items-center gap-6">
-        <div className="col-span-6">
-          <div className="flex items-center gap-5">
-            <div className="flex-shrink-0">
-              <Image
-                src={image}
-                alt={title}
-                width={96}
-                height={64}
-                className="h-30 w-40 object-cover"
-              />
-            </div>
-
-            {/* Course Info */}
-            <div className="min-w-0 flex-1">
-              {/* Rating */}
-              <div className="mb-2 flex items-center gap-1">
-                <Icon icon="ph:star-fill" className="text-primary text-sm" />
-                <span className="text-base-content/70 text-xs font-semibold">
-                  {rating} ({reviews.toLocaleString()} Review)
-                </span>
-              </div>
-
-              {/* Title */}
-              <h3 className="text-base-content/80 mb-2 text-sm font-medium">
-                {title}
-              </h3>
-
-              {/* Instructors */}
-              <span className="text-base-content/70 text-xs">
-                <span className="text-base-content/50"> Course by: </span>
-                {instructors}
-              </span>
-            </div>
-          </div>
-        </div>
-
-        {/* Prices Column */}
-        <div className="col-span-3 text-center">
-          <div className="flex flex-row items-center gap-1">
-            <span className="text-primary text-xl font-medium">{price}</span>
-            {originalPrice && (
-              <span className="text-base-content/40 text-xs line-through">
-                {originalPrice}
-              </span>
-            )}
-          </div>
-        </div>
-
-        {/* Action Column */}
-        <div className="col-span-2 ml-8 text-center text-nowrap">
-          <div className="flex items-center justify-center gap-3">
-            <form method="POST" action="/api/cart/buy-now">
-              <input type="hidden" name="courseId" value={id} />
-              <button
-                type="submit"
-                className="btn btn-soft btn-base-content h-10 w-28 text-sm"
-              >
-                Buy Now
-              </button>
-            </form>
-
-            <form method="POST" action="/api/cart/add">
-              <input type="hidden" name="courseId" value={id} />
-              <button
-                type="submit"
-                className="btn btn-primary h-10 w-28 text-sm"
-              >
-                Add To Cart
-              </button>
-            </form>
-
-            <form method="POST" action="/api/wishlist/remove">
-              <input type="hidden" name="courseId" value={id} />
-              <button
-                type="submit"
-                className="btn btn-soft btn-primary h-8 w-8 p-0"
-                title="Remove from wishlist"
-              >
-                <Icon icon="ph:heart-fill" className="text-sm" />
-              </button>
-            </form>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-};
